% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clustering.R
\name{cluster_pairwise_effects}
\alias{cluster_pairwise_effects}
\title{Cluster by Causal Effects}
\usage{
cluster_pairwise_effects(results, pairwise_effects, k, cluster_method,
  hclust_method, dist_measure, alpha = 0.95, iterations_pv, protein, outpath,
  file_separator)
}
\description{
Cluster by Causal Effects
}
