% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graphs.R
\name{conflict_edges}
\alias{conflict_edges}
\title{Edges Types in Graph
calls edge_information with full adj_matrix if interesting_positions is missing or the rows and columns of that vector if it is given}
\usage{
conflict_edges(graph, print = FALSE, interesting_positions)
}
\description{
Edges Types in Graph
calls edge_information with full adj_matrix if interesting_positions is missing or the rows and columns of that vector if it is given
}
