% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graphs.R
\name{edge_information}
\alias{edge_information}
\title{Different Edge Types in a Graph}
\usage{
edge_information(adj_m, print = FALSE)
}
\arguments{
\item{adj_m}{the adjacency matrix of the graph that is to be considered (as returned by wgtMartix(...))}

\item{print}{should the result also be printed?}
}
\value{
A list with names "conflict", "directed" and "undirected" and the respective numbers of such edges.
}
\description{
Get the number of edges of different types (conflict, bidirected, directed)
}
