% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graphs.R
\name{enumerate_graphs_rec}
\alias{enumerate_graphs_rec}
\title{Given a graph with some edges direction unclear, this function enumerates all possible graphs by fixing those unclear edges in every possible direction.}
\usage{
enumerate_graphs_rec(am, n, i = 1, j = 1)
}
\arguments{
\item{am}{A adjecency matrix where 0 indicates no edge, 1 an edge, and 2 an edge with unclear direction.}

\item{n}{Number of nodes.}

\item{i}{Parameter needed for recursion, set to 1 at the initial call.}

\item{j}{Parameter needed for recursion, set to 1 at the initial call.}
}
\description{
Given a graph with some edges direction unclear, this function enumerates all possible graphs by fixing those unclear edges in every possible direction.
}
