% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_causal_effects.R
\name{get_most_influenced_positions}
\alias{get_most_influenced_positions}
\title{Positions with highest effects}
\usage{
get_most_influenced_positions(effects, threshold, percentile)
}
\arguments{
\item{effects}{Vector (or matrix) of effects that are to be considered}

\item{threshold}{threshold above which effects are considered high}

\item{percentile}{percentile of highest positions to be returned}
}
\value{
A vector (?) of positions whose effects are above the \code{threshold} or in the top \code{percentile}.
}
\description{
Returns the labels of the positions for which the effects are highest
}
\details{
Either \code{threshold} or \code{percentile} must be given.
If both are given, percentile is ignored.
}
