% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graphs.R
\name{kernelize_graph}
\alias{kernelize_graph}
\title{Removes isolated nodes from a graph.}
\usage{
kernelize_graph(graph, ret_list = FALSE)
}
\arguments{
\item{graph}{A graph object as described in the R graph package.}

\item{ret_list}{A boolean flag indicating if the removed nodes should be returned.}
}
\value{
The kernelized graph and if ret_list is the additionally the list of removed nodes.
}
\description{
Removes isolated nodes from a graph.
}
