% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_clustering.R
\name{protein_graph_clustering}
\alias{protein_graph_clustering}
\title{Cluster by Causal Structure (Graph Clustering)}
\usage{
protein_graph_clustering(results, protein, position_numbering, coloring, colors,
  outpath, output_formats, file_separator, mute_all_plots, caption,
  cluster_methods, add_cluster_of_conserved_positions, removed_cols,
  more_levels_of_conservedness = FALSE, sort_clusters = length)
}
\description{
Cluster by Causal Structure (Graph Clustering)
}
