% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tools.R
\name{remove_zero_rows_or_columns}
\alias{remove_zero_rows_or_columns}
\title{Remove the rows or columns in a matrix that only contain zeros}
\usage{
remove_zero_rows_or_columns(matrix, rows_or_columns)
}
\arguments{
\item{matrix}{the matrix}

\item{rows_or_columns}{if set to 1, zero rows are removed, otherwise zero columns}
}
\description{
Remove the rows or columns in a matrix that only contain zeros
}
