% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tools.R
\name{reorder_list_of_lists}
\alias{reorder_list_of_lists}
\title{This function is used e.g. for plotting connected components in pymol.
Normally, the connected components are sorted alphabetically.
Since in proteins the false impression that neighbouring residues are in the
same connected components could arise when they have subsequent and thus
indistinguishable rainbow colors, the components can be mixed (mix argment given).
Then, mix_offset determines how many other positions there are (at least) between two
originally neighbouring components}
\usage{
reorder_list_of_lists(list, ordering,
  mix_mode = "mix_offset_between_original_neighbours",
  mix_offset = floor(sqrt(length(list))), sort_mode, sort_descending = TRUE)
}
\arguments{
\item{mix}{if mix = "every_mix_offset_th", the mixed version is obtained by appending to the
new list every i'th element from the original list, where i = mix_offset + 1,
continuing with incremeted offset every time the end of the list is reached}
}
\description{
This function is used e.g. for plotting connected components in pymol.
Normally, the connected components are sorted alphabetically.
Since in proteins the false impression that neighbouring residues are in the
same connected components could arise when they have subsequent and thus
indistinguishable rainbow colors, the components can be mixed (mix argment given).
Then, mix_offset determines how many other positions there are (at least) between two
originally neighbouring components
}
